Fix code to run FCN locally.

+++ mlsp_final/metrics.py
@@ -7,13 +7,11 @@ import numpy as np
 class runningScore(object):
     def __init__(self, n_classes):
         self.n_classes = n_classes
-        self.confusion_matrix = np.zeros((n_classes, n_classes))
+        self.confusion_matrix = np.zeros((n_classes, n_classes)).astype(np.int32)
 
     def _fast_hist(self, label_true, label_pred, n_class):
         mask = (label_true >= 0) & (label_true < n_class)
-        hist = np.bincount(
-            n_class * label_true[mask].astype(int) + label_pred[mask], minlength=n_class ** 2
-        ).reshape(n_class, n_class)
+        hist = np.bincount(n_class * label_true[mask].astype(int) + label_pred[mask], minlength=n_class ** 2).reshape(n_class, n_class)
         return hist
 
     def update(self, label_trues, label_preds):
